requirements.txt
python-dotenv
smtplib
email
glob